set up route handlers for two different requests from client side, ALSO created a flightController folder as well as two middlewares for routing different get requests from the client
